+++ BlogposterCMS/tests/placeholderParityCheck.js
@@ -1,0 +1,2 @@
+require('./sqlitePlaceholderParity.test');
+require('./mongoPlaceholderParity.test');